hasura-specific-provisioner/pyproject.toml
[tool.poetry]
name = "hasura-specific-provisioner"
version = "0.1.0"
description = "Hasura Specific Provisioner for Witboost"
license = "Apache-2.0"
authors = ["Nicolò Bidotti <nicolo.bidotti@agilelab.it>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
black = "^23.9.1"
fastapi = {extras = ["all"], version = "^0.103.1"}
httpx = "^0.25.0"
mypy = "^1.5.1"
opentelemetry-distro = "^0.41b0"
opentelemetry-exporter-otlp-proto-grpc = "^1.20.0"
opentelemetry-exporter-otlp = "^1.20.0"
opentelemetry-instrumentation-fastapi = "0.41b0"
pip-audit = "^2.6.1"
pre-commit = "^3.4.0"
pytest = "^7.4.2"
pytest-cov = "^4.0.0"
ruff = "^0.0.291"
pyyaml = "^6.0.1"
types-pyyaml = "^6.0.1"
types-requests = "^2.31.0.5"
urllib3 = "^2.0.6"

[tool.poetry.group.dev.dependencies]
testcontainers = "^3.7.1"

[tool.ruff]
select = ["E", "F", "I"]

[tool.black]
skip-string-normalization = true

[tool.pytest.ini_options]
addopts = "-v"
log_cli = true
log_cli_level = "DEBUG"

[tool.coverage.report]
fail_under = 85

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.liccheck]
authorized_licenses = [
    'apache license 2.0',
    'apache software',
    'bsd',
    'cc0 1.0 universal (cc0 1.0) public domain dedication',
    'gnu lesser general public license v2 or later (lgplv2+)',
    'isc',
    'isc license (iscl)',
    'mit',
    'mozilla public license 2.0 (mpl 2.0)',
    'other/proprietary',
    'python software foundation',
    'the unlicense (unlicense)',
    '3-clause bsd',
    'BSD-2-Clause',
    'Apache-2.0'
]
unauthorized_licenses = [
    "gnu general public license v2 (gplv2)"
]
